+++ javageci-core/src/main/java/javax0/geci/fluent/Fluent.java
@@ -11,7 +11,7 @@ import javax0.geci.log.Logger;
 import javax0.geci.log.LoggerFactory;
 import javax0.geci.tools.AbstractGenerator;
 import javax0.geci.tools.CompoundParams;
-import javax0.geci.tools.Tools;
+import javax0.geci.tools.GeciReflectionTools;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
@@ -49,7 +49,7 @@ public class Fluent extends AbstractGenerator {
         var methodName = s.substring(sepPos + 2);
         final Class<?> klass;
         try {
-            klass = Tools.classForName(className);
+            klass = GeciReflectionTools.classForName(className);
         } catch (ClassNotFoundException e) {
             throw new GeciException("definedBy class '" + className + "' can not be found");
         }
