+++ javageci-core/src/main/java/javax0/geci/fluent/internal/FluentMethodTool.java
@@ -1,7 +1,7 @@
 package javax0.geci.fluent.internal;
 
 import javax0.geci.tools.MethodTool;
-import javax0.geci.tools.Tools;
+import javax0.geci.tools.GeciReflectionTools;
 
 import java.lang.reflect.Type;
 
@@ -15,12 +15,12 @@ public class FluentMethodTool extends MethodTool {
 
     private FluentMethodTool(Class klass) {
         super();
-        this.klassName = Tools.getGenericTypeName(klass);
+        this.klassName = GeciReflectionTools.getGenericTypeName(klass);
     }
 
     @Override
     public String getArgCall(Type t) {
-        final var normType = Tools.getGenericTypeName(t);
+        final var normType = GeciReflectionTools.getGenericTypeName(t);
         final String arg;
         if (normType.equals(klassName)) {
             arg = "((Wrapper)arg" + argCounter.addAndGet(1) + ").that";
@@ -32,7 +32,7 @@ public class FluentMethodTool extends MethodTool {
 
     @Override
     public String getArg(Type t) {
-        final var normType = Tools.getGenericTypeName(t);
+        final var normType = GeciReflectionTools.getGenericTypeName(t);
         final String actualType;
         if (normType.equals(klassName)) {
             actualType = "WrapperInterface";
